+++ src/solvers/hidden_tuples.rs
@@ -20,12 +20,12 @@ pub fn find_hidden_tuples(board: &Board, size: usize, strategy: Strategy) -> Opt
         Known::ALL
             .into_iter()
             .map(|k| (k, house.cells() & board.candidate_cells(k)))
-            .filter(|(_, candidates)| 2 <= candidates.size() && candidates.size() as usize <= size)
+            .filter(|(_, candidates)| 2 <= candidates.size() && candidates.size() <= size)
             .combinations(size)
             .for_each(|candidates| {
                 let cell_sets = candidates.iter().map(|(_, cs)| *cs).collect::<Vec<_>>();
                 let cells = cell_sets.iter().copied().union() as CellSet;
-                if cells.size() as usize != size
+                if cells.size() != size
                     || is_degenerate(&cell_sets, size, 2)
                     || is_degenerate(&cell_sets, size, 3)
                 {
@@ -59,7 +59,7 @@ fn is_degenerate(cell_sets: &[CellSet], size: usize, smaller_size: usize) -> boo
             .iter()
             .combinations(smaller_size)
             .map(|sets| sets.into_iter().copied().union())
-            .any(|set| (set.size() as usize) <= smaller_size)
+            .any(|set| (set.size()) <= smaller_size)
 }
 
 #[cfg(test)]
